move over #input to new doc structure
