+++ packages/dnb-design-system-portal/src/pages/uilib/components/input-masked.md
@@ -5,11 +5,22 @@ status: null
 order: 7
 ---
 
-<!--
-  ATTENTION: This file is auto generated by using "makeDemosFactory".
-  Do not change the content!
--->
+import Tabs from 'Tags/Tabs'
 
-import InputMaskedPage from 'Src/uilib/components/demos/InputMasked'
+import InputMaskedInfo from 'Pages/uilib/components/input-masked/input-masked-info'
+import InputMaskedProperties from 'Pages/uilib/components/input-masked/input-masked-properties'
+import InputMaskedEvents from 'Pages/uilib/components/input-masked/input-masked-events'
 
-<InputMaskedPage />
+# InputMasked
+
+<Tabs>
+  <Tabs.Content>
+    <InputMaskedInfo />
+  </Tabs.Content>
+  <Tabs.Content>
+    <InputMaskedProperties />
+  </Tabs.Content>
+  <Tabs.Content>
+    <InputMaskedEvents  />
+  </Tabs.Content>
+</Tabs>
