+++ packages/dnb-design-system-portal/src/pages/uilib/components/input-masked/input-masked-events.md
@@ -1,0 +1,9 @@
+---
+draft: true
+---
+
+| Events            | Description                                                                 |
+| ----------------- | --------------------------------------------------------------------------- |
+| `on_change`       | _(optional)_ will be called on state changes made by the user.              |
+| `on_submit`       | _(optional)_ will be called on submit button click.                         |
+| `on_state_update` | _(optional)_ will be called once the parameter `checked` changes its value. |
