+++ packages/dnb-design-system-portal/src/pages/uilib/components/input-masked/input-masked-info.md
@@ -1,5 +1,15 @@
+---
+draft: true
+---
+
+import Examples from 'Pages/uilib/components/input/Examples'
+
 ## Description
 
 The masked input component uses the DNB original input component, but with some additional masking functionality. Read more about the details [on the open source project](github.com/sanniassin/react-input-mask).
 
 **TODO:** This is an experimental component and should be tested more before using it in production (thinking of e.g. DerivedStateFromProps, or that the "mask" prop uses a function or RegExp, so there has to be build in a special type of JSON parser).
+
+## Demos
+
+<Examples />
