+++ packages/dnb-design-system-portal/src/pages/uilib/components/input.md
@@ -5,11 +5,22 @@ status: null
 order: 8
 ---
 
-<!--
-  ATTENTION: This file is auto generated by using "makeDemosFactory".
-  Do not change the content!
--->
+import Tabs from 'Tags/Tabs'
 
-import InputPage from 'Src/uilib/components/demos/Input'
+import InputInfo from 'Pages/uilib/components/input/input-info'
+import InputProperties from 'Pages/uilib/components/input/input-properties'
+import InputEvents from 'Pages/uilib/components/input/input-events'
 
-<InputPage />
+# Input
+
+<Tabs>
+  <Tabs.Content>
+    <InputInfo />
+  </Tabs.Content>
+  <Tabs.Content>
+    <InputProperties />
+  </Tabs.Content>
+  <Tabs.Content>
+    <InputEvents  />
+  </Tabs.Content>
+</Tabs>
