+++ packages/dnb-design-system-portal/src/pages/uilib/components/input/input-info.md
@@ -1,3 +1,9 @@
+---
+draft: true
+---
+
+import Examples from 'Pages/uilib/components/input/Examples'
+
 ## Description
 
 The input component is an umbrella component for all inputs which share the same style as the classic `text` input field. Radio buttons and other form elements are not included here.
@@ -5,3 +11,7 @@ The input component is an umbrella component for all inputs which share the same
 There is an experimental [Masked-Input component](/uilib/components/input-masked/)
 
 The **id** property is mandatory because of it's importance in the DOM - whether it's as a reference for the label or as a hook when submitting a form.
+
+## Demos
+
+<Examples />
