--- packages/dnb-design-system-portal/src/uilib/components/demos/Input.js
@@ -1,44 +1,0 @@
-/**
- * ATTENTION: This file is auto generated by using "makeDemosFactory".
- * Do not change the content!
- *
- * "Component" Demo setup
- * Ready for imporing in page
- *
- */
-
-import React, { PureComponent } from 'react'
-import PropTypes from 'prop-types'
-import ItemWrapper from '../../../shared/parts/uilib/ItemWrapper'
-import DemoComponent, {
-  Example as CodeComponent
-} from 'dnb-ui-lib/src/components/input/Example'
-import Description from 'dnb-ui-lib/src/components/input/description.md'
-import Details from 'dnb-ui-lib/src/components/input/details.md'
-import ExampleCode from 'raw-loader!../examples/Input.txt'
-
-export default class InputDemo extends PureComponent {
-  static propTypes = {
-    title: PropTypes.string,
-    id: PropTypes.string,
-    Description: PropTypes.func,
-    Details: PropTypes.func,
-    DemoComponent: PropTypes.func,
-    CodeComponent: PropTypes.func,
-    hideTabs: PropTypes.bool
-  }
-  static defaultProps = {
-    title: 'Input',
-    id: 'input',
-    ExampleCode,
-    Description,
-    Details,
-    DemoComponent,
-    CodeComponent,
-    callback: CodeComponent.AdditionalCallback || null,
-    hideTabs: false
-  }
-  render() {
-    return <ItemWrapper {...InputDemo.defaultProps} {...this.props} />
-  }
-}
