+++ packages/dnb-design-system-portal/src/uilib/components/demos/allParts.js
@@ -9,8 +9,8 @@ import FormLabel from './FormLabel'
 import FormStatus from './FormStatus'
 import Icon from './Icon'
 import IconPrimary from './IconPrimary'
-import Input from './Input'
-import InputMasked from './InputMasked'
+// import Input from './Input'
+// import InputMasked from './InputMasked'
 import LineTitle from './LineTitle'
 import Logo from './Logo'
 import Modal from './Modal'
@@ -27,8 +27,8 @@ export default [
   FormStatus,
   Icon,
   IconPrimary,
-  Input,
-  InputMasked,
+  // Input,
+  // InputMasked,
   LineTitle,
   Logo,
   Modal,
